+++ sea-query-derive/tests/compile-fail/enum_def/enum.rs
@@ -1,0 +1,8 @@
+use sea_query_derive::enum_def;
+
+#[enum_def]
+enum Hello {
+    Name,
+}
+
+fn main() {}