+++ sea-query-derive/tests/pass/enum_def/default.rs
@@ -1,0 +1,10 @@
+use sea_query_derive::enum_def;
+
+#[enum_def]
+pub struct Hello {
+    pub name: String
+}
+
+fn main() {
+    assert_eq!(format!("{:?}", HelloIden::Name), "Name");
+}