+++ src/index/create.rs
@@ -1,6 +1,7 @@
 use inherent::inherent;
 
 use crate::{backend::SchemaBuilder, types::*, SchemaStatementBuilder};
+use crate::{ConditionHolder, ConditionalStatement, IntoCondition};
 
 use super::common::*;
 
@@ -148,6 +149,27 @@ use super::common::*;
 ///     r#"CREATE INDEX "idx-glyph-aspect" ON "glyph" ("aspect" ASC)"#
 /// );
 /// ```
+///
+/// Partial Index with prefix and order
+/// ```
+/// use sea_query::{tests_cfg::*, *};
+///
+/// let index = Index::create()
+///     .name("idx-glyph-aspect")
+///     .table(Glyph::Table)
+///     .col((Glyph::Aspect, 64, IndexOrder::Asc))
+///     .and_where(Expr::col((Glyph::Table, Glyph::Aspect)).is_in(vec![3, 4]))
+///     .to_owned();
+///
+/// assert_eq!(
+///     index.to_string(PostgresQueryBuilder),
+///     r#"CREATE INDEX "idx-glyph-aspect" ON "glyph" ("aspect" (64) ASC) WHERE "glyph"."aspect" IN (3, 4)"#
+/// );
+/// assert_eq!(
+///     index.to_string(SqliteQueryBuilder),
+///     r#"CREATE INDEX "idx-glyph-aspect" ON "glyph" ("aspect" ASC) WHERE "glyph"."aspect" IN (3, 4)"#
+/// );
+/// ```
 #[derive(Default, Debug, Clone)]
 pub struct IndexCreateStatement {
     pub(crate) table: Option<TableRef>,
@@ -157,6 +179,7 @@ pub struct IndexCreateStatement {
     pub(crate) nulls_not_distinct: bool,
     pub(crate) index_type: Option<IndexType>,
     pub(crate) if_not_exists: bool,
+    pub(crate) r#where: ConditionHolder,
 }
 
 /// Specification of a table index
@@ -171,7 +194,16 @@ pub enum IndexType {
 impl IndexCreateStatement {
     /// Construct a new [`IndexCreateStatement`]
     pub fn new() -> Self {
-        Self::default()
+        Self {
+            table: None,
+            index: Default::default(),
+            primary: false,
+            unique: false,
+            nulls_not_distinct: false,
+            index_type: None,
+            if_not_exists: false,
+            r#where: ConditionHolder::new(),
+        }
     }
 
     /// Create index if index not exists
@@ -263,6 +295,7 @@ impl IndexCreateStatement {
             nulls_not_distinct: self.nulls_not_distinct,
             index_type: self.index_type.take(),
             if_not_exists: self.if_not_exists,
+            r#where: self.r#where.clone(),
         }
     }
 }
@@ -282,4 +315,19 @@ impl SchemaStatementBuilder for IndexCreateStatement {
     }
 
     pub fn to_string<T: SchemaBuilder>(&self, schema_builder: T) -> String;
+}
+
+impl ConditionalStatement for IndexCreateStatement {
+    fn and_or_where(&mut self, condition: LogicalChainOper) -> &mut Self {
+        self.r#where.add_and_or(condition);
+        self
+    }
+
+    fn cond_where<C>(&mut self, condition: C) -> &mut Self
+    where
+        C: IntoCondition,
+    {
+        self.r#where.add_condition(condition.into_condition());
+        self
+    }
 }